Bugfixes and it's working
1. Checking if the session exists needs double brackets for the conditional, as
well as piping stderr to dev/null. `tmux` uses stderr to complain if it can't
find the session. Based on the man page, there appears to be no way to turn
this off.
2. The check on the GOTO_PROJECT_DIR variable needs to be short-circuited with
an &&, not an ||, so that HOME is used as the default only if GOTO_PROJECT_DIR
is empty. In the *very* strange case that you want to put all your projects in
"/", you should explicitly set that.

TODO: completions